Make Free Memory menu safe.  Not perfect yet
This fixes some crash bugs on D678:
 - SRM buffers crash after a delay if you try to allocate more than 4
 - Trying to create a task with a large requested stack size is an immediate crash

I restrict stack size usage by bin_search() to a safe value for 200D, 72kB.
Not tested on other cams.  This is also not the true free stack space,
it will always report 71 for me, so, not ideal.

We should work out how to determine free stack space properly.  The approach
on old cams isn't great but does work there, surely there is a more direct
way that works on both new and old.

The SRM buffer fix is cleaner, can probably be kept unchanged.