D678: more tidying boot-d678.c, allow thumb or ARM D6 builds
Make patch_thumb_branch() auto-detect whether to BL or BLX.
This simplifies logic elsewhere.

It also allows D6 builds to use -mthumb in Makefile.platform.default.
Not enabled this yet as I don't know why ARM was chosen (maybe
Alex simply defaulted to this, since D45 used it and D6 worked with it?)

Thumb saves 65k from the ML binary so it's nice if we can use it.