5D4: remove bad magic for starting second digic
Other dual digic 6 cams seem to wake up the second digic
with a magic value sent in early code.  But 5D4 doesn't do
this, and polling for it hung forever.

This now boots to Canon menus.  Don't know if the secondary
digic is working, I assume so since forums suggest it would
WDT kill us otherwise.