650d: possible fix for "isoless err" by removing magic numbers
Old code use a fixed heap address and assumed it would never change,
this was not true.

New code searches the heap structs (at least, I think that's what they
are...) to find what I think is a stored DMA struct used for the
transfer that copies the ISO data from serial flash to ram.

See ff122548() which allocates the heap and assigns to the struct,
from here you can see why I compare which fields in
get_photo_cmos_iso_start_650d()

There are some differences in the DMA struct thing compared to 550d,
but they're quite similar.  We might be able to generalise, and turn
the offset between the FRAME and PHOTO STARTs into a per cam constant.