task structs: make new types required
This completes the work to make each cam define which
task and task_attr structs the OS uses.  There is no longer
a default.

Modules no longer have a definition for the structs so attempts
to use fields will fail.  This is good: previously, a module would
always use the default structs, and the offsets of fields would get
compiled in.  Modules are supposed to be cross-cam compatible, but
this would not work.