modules: save cross-module deps into modules
We now add a .module_deps section to any module that
depends on exports from another module.

The section contains an array of null-terminated strings,
each string being a module name.  Each such module contains
one or more exports this module requires in order to load.

The intent is this will be used to automatically load
modules at runtime on cam.