modules: fix UB around bad func pointer casts
We were relying on undefined behaviour with some
func pointer casts in the module CBR code.  Using a union
makes it explicit that we intend to mangle two different
types of function into this code.

I don't like this pattern, I think the CBR code should
be reworked.  CBR_NAMED seems to be used in a different
enough way from the old CBRs that maybe it should be split out?

This change removes the UB and shouldn't change behaviour at all,
so it's a strict improvement (I think!).

In general I'd say this CBR code feels overengineered.
Look at the arena stuff, do we ever even hit this code?