200D: working EDMAC mem->mem copies (demo)
Yeah, that's right, we got EDMAC now.

This is quite limited, as my understanding is limited.
There's a nice function in 200D that does everything in one place,
but hard-codes a bunch of stuff so it's only for mem->mem copies.
That's the new stub.

I expect this is the break we need to gain full understanding
of the EDMAC changes.  It's now absolutely proven EDMAC exists
and works on 200D.

See long comments in debug.c