1100D: remove features to make build small enough
I don't care much about the 1100D so this is a quick fix.
Slightly oddly, the build getting too big to boot was reported before:
https://github.com/reticulatedpines/magiclantern_simplified/issues/109

I fixed that, so something made the build quite a lot larger since.
I think I upgrade gcc in that window, so that's worth testing.
It would be odd for that to cause a large size increase since
we build most parts with -Os