Disable a bunch of features in order to save autoexec.bin space
Otherwise, cam would crash at boot because it places BSS outside the space available
for ML (ml_reserved_mem).

BSS is placed next to autoexec.bin, and (last time I checked) it assumed the space would be available
with no checks.

My memory isn't that great from this, and I think I misread a lot. So check this for the full story:

https://discord.com/channels/671072748985909258/807566022424789022/1279547942989987997